- Removed a header from www - Added modular email function to test - Created emailfunctions.js for all emails we will send
